+++ src/server/devices/sharedFunctions/index.ts
@@ -1,1 +1,2 @@
 export * from './lampUtils';
+export * from './temperatureSensorService';