+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ## **WORK IN PROGRESS**
 
 * (theimo1221) Allow Heating Rod even during winter
+* (theimo1221) Generalize Temperature Sensor
 
 ## 3.0.0-alpha.89 (2024-10-03)